Reduce Java agent to just instrumenter and asm packages.
