+++ quasar-core/src/main/java/co/paralleluniverse/common/util/ExtendedStackTraceElement.java
@@ -12,7 +12,6 @@
  */
 package co.paralleluniverse.common.util;
 
-// import java.lang.reflect.Executable;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Member;
 import java.lang.reflect.Method;
