+++ quasar-core/src/main/java/co/paralleluniverse/strands/channels/Selector.java
@@ -13,10 +13,10 @@
  */
 package co.paralleluniverse.strands.channels;
 
-import co.paralleluniverse.common.monitoring.Debug;
 import co.paralleluniverse.common.monitoring.FlightRecorder;
 import co.paralleluniverse.common.monitoring.FlightRecorderMessage;
 import co.paralleluniverse.common.reflection.GetDeclaredField;
+import co.paralleluniverse.common.util.Debug;
 import co.paralleluniverse.common.util.UtilUnsafe;
 import co.paralleluniverse.fibers.SuspendExecution;
 import co.paralleluniverse.strands.Strand;
